Ajout fichier fonctionsMaths.js
