+++ src/js/exercices/3e/3G30-MG32.js
@@ -60,7 +60,7 @@ export default function Exercice_Trigo_longueurs() {
     ]);
     let k1 = Math.round((Math.random() * 5 + 1) * 10) / 10;
     let k2 = Math.round((Math.random() * 5 + 1) * 10) / 10;
-    angle1 = Math.round(Math.degres(Math.atan(k2 / k1)));
+    angle1 = Math.round(degres(Math.atan(k2 / k1)));
     let alpha1 = Math.random() * Math.PI - Math.PI / 2;
     let alpha1deg = Math.round((alpha1 * 180) / Math.PI);
     let x1 = k1; // coordonnées des deux sommets du triangle
