+++ src/js/modules/outils.js
@@ -866,62 +866,6 @@ export function valeurBase (n) {
 }
 
 /**
-* Convertit un angle de radian vers degrés et fonction inverse
-* @Example
-* // PI->180
-* @Auteur Jean-Claude Lhote
-*/
-Math.degres = function (radians) {
-  return radians * 180 / Math.PI
-}
-Math.radians = function (degres) {
-  return degres * Math.PI / 180
-}
-
-/**
- * @param {number} a angle en degrés
- * @returns flottant : le cosinus de l'angle
- */
-export function cos (a) {
-  return calcul(Math.cos(Math.radians(a)))
-}
-/**
- * @param {number} a angle en degrés
- * @returns flottant : le sinus de l'angle
- */
-export function sin (a) {
-  return calcul(Math.sin(Math.radians(a)))
-}
-/**
- * @param {number} a angle en degrés
- * @returns flottant : la tangente de l'angle
- */
-export function tan (a) {
-  return calcul(Math.tan(Math.radians(a)))
-}
-/**
- * @param {number} un nombre qui correspond au cosinus de l'angle
- * @returns flottant : la mesure de l'angle en degrés
- */
-export function ArcCos (x) {
-  return arrondi(Math.degres(Math.acos(x)), 1)
-}
-/**
- * @param {number} un nombre qui correspond au sinus de l'angle
- * @returns flottant : la mesure de l'angle en degrés
- */
-export function ArcSin (x) {
-  return arrondi(Math.degres(Math.asin(x)), 1)
-}
-/**
- * @param {number} un nombre qui correspond à la tangente de l'angle
- * @returns flottant : la mesure de l'angle en degrés
- */
-export function ArcTan (x) {
-  return arrondi(Math.degres(Math.atan(x)), 1)
-}
-
-/**
  *
  * @param {array} matrice M tableau 3x3 nombres
  * @param {array} vecteur A tableau 3 nombres
